deploy a less destructive lambda for now
